sync.c: avoid printf format warning
* exec/sync.c (sync_deliver_fn): cast to unsigned long int, use %lu

git-svn-id: http://svn.fedorahosted.org/svn/corosync/trunk@2034 fd59a12c-fef9-0310-b244-a6a79926bd2f

+++ exec/sync.c
@@ -336,7 +336,8 @@ static void sync_deliver_fn (
 	unsigned int barrier_completed;
 	int i;
 
-log_printf (LOG_LEVEL_DEBUG, "confchg entries %d\n", barrier_data_confchg_entries);
+	log_printf (LOG_LEVEL_DEBUG, "confchg entries %lu\n",
+		    (unsigned long int) barrier_data_confchg_entries);
 	if (endian_conversion_required) {
 		sync_endian_convert (req_exec_sync_barrier_start);
 	}
